Fixed json_extractor to include null values in extractor input
We used to intentionally eliminate fields in a row where value
is null in JSON extractor input. Now we want to include these
fields with null values.

For example, for a table a(x, y), if a row is (1, NULL), the
json_extractor input for this row used to be {"x": 1}. Now our input is
{"x": 1, "y": null}.